Added data Models for Team and TeamMember
I implemented the data models for Team and TeamMember. I've also updated the Todo model to reference the a Team foreign key and I think I got the create form working with that. The admin panel also works now to view the data models. Make sure to run the migrations and you might also need to just delete the DB for it to populate again.