Fix categories menu not showing categories from team todos
Previously, the categories drop down only displayed options from items that are directly created by the user. I've fixed the logic so now it will show categories from all the todos the user is in as well.

The logic had to get a little gross here to keep a unfiltered copy of the todos because the order of operations in django doesn't allowing filters after a union.